requirements.txt
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.5
crashtest==0.4.1
distlib==0.3.7
dulwich==0.21.5
filelock==3.12.2
frozenlist==1.4.0
graphql-core==3.2.3
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
keyring==23.13.1
libcst==1.0.1
lockfile==0.12.2
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.2.0
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.15.1
pyproject_hooks==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0
rapidfuzz==2.15.1
referencing==0.29.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
rpds-py==0.8.11
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
starlette==0.30.0
strawberry-graphql==0.194.4
tomlkit==0.11.8
trove-classifiers==2023.7.6
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.21.1
virtualenv==20.24.0
webencodings==0.5.1
xattr==0.10.1
yarl==1.9.2
zipp==3.16.2
